Extracted a RequestFactory interface
